Cleanup at end of feature.
